refactor(react-gui): separate the render settings' arithmetic and its hatch editor
useRenderSettings was 534 lines holding three unrelated things: the
arithmetic on the settings list, a layer editor for the NPR hatch look,
and the settings state itself.

The settings are a flat list of `{ key, value }` rows rather than a typed
object, because each backend declares its own. That makes every read and
write a search and unit conversion a walk over the list -- 119 lines of
list arithmetic with no React in it, which now sits in propMath.ts where
it can be read without a hook.

The hatch editor is a separate piece of state with its own rule: the
selected style is a template, the editor holds an editable copy, and
`hatchDirty` decides whether the look travels with the render at all --
an untouched one is not sent, so the render uses the style's own
configuration rather than a round trip. It takes `hatchStyle` as an
argument rather than reading the backend props itself, because two
readers of that value would be two ways for them to disagree.

Three exhaustive-deps omissions surfaced on the way, all the same shape:
`setHatch` was a closure capture and is now an argument, so the three
callbacks that reset the edited look have to name it. Declaring the
editor above them was the ordering that made that possible.
